feat: move utils to internal
