+++ compose/graph_node.go
@@ -22,7 +22,7 @@ import (
 	"reflect"
 
 	"github.com/cloudwego/eino/components"
-	"github.com/cloudwego/eino/utils/generic"
+	"github.com/cloudwego/eino/internal/generic"
 )
 
 // the info of most original executable object directly provided by the user
