+++ internal/callbacks/inject.go
@@ -19,8 +19,8 @@ package callbacks
 import (
 	"context"
 
+	"github.com/cloudwego/eino/internal/generic"
 	"github.com/cloudwego/eino/schema"
-	"github.com/cloudwego/eino/utils/generic"
 )
 
 type CtxManagerKey struct{}
