+++ internal/generic/generic.go
