+++ internal/safe/panic.go
